app-engine-api-spotify-server/requirements.txt
Flask==1.0.2
flask-cors==3.0.7
gunicorn==19.9.0
six==1.11.0
pyyaml>=4.2b1
requests==2.20.1
google-auth==1.6.1
google-auth-oauthlib==0.2.0
google-cloud-storage
google-cloud-datastore
grpcio